Moved account renaming to `LocalTaskListStore` and trying to fix issue
