+++ app/src/main/kotlin/at/bitfire/davdroid/repository/AccountRepository.kt
@@ -14,7 +14,7 @@ import at.bitfire.davdroid.db.Credentials
 import at.bitfire.davdroid.db.HomeSet
 import at.bitfire.davdroid.db.Service
 import at.bitfire.davdroid.resource.LocalAddressBookStore
-import at.bitfire.davdroid.resource.LocalTaskList
+import at.bitfire.davdroid.resource.LocalTaskListStore
 import at.bitfire.davdroid.servicedetection.DavResourceFinder
 import at.bitfire.davdroid.servicedetection.RefreshCollectionsWorker
 import at.bitfire.davdroid.settings.AccountSettings
@@ -49,6 +49,7 @@ class AccountRepository @Inject constructor(
     private val collectionRepository: DavCollectionRepository,
     private val homeSetRepository: DavHomeSetRepository,
     private val localAddressBookStore: Lazy<LocalAddressBookStore>,
+    private val localTaskListStoreFactory: LocalTaskListStore.Factory,
     private val logger: Logger,
     private val settingsManager: SettingsManager,
     private val serviceRepository: DavServiceRepository,
@@ -223,7 +224,10 @@ class AccountRepository @Inject constructor(
 
             // update account_name of local tasks
             try {
-                LocalTaskList.onRenameAccount(context, oldAccount.name, newName)
+                TaskProvider.acquire(context)?.use { provider ->
+                    val dataStore = localTaskListStoreFactory.create(provider.name.authority)
+                    dataStore.renameAccount(provider.client, oldAccount.name, newName)
+                }
             } catch (e: Exception) {
                 logger.log(Level.WARNING, "Couldn't propagate new account name to tasks provider", e)
                 // Couldn't update task lists, but this is not a fatal error (will be fixed at next sync)
