+++ app/src/main/kotlin/at/bitfire/davdroid/resource/LocalTaskList.kt
@@ -5,10 +5,8 @@
 package at.bitfire.davdroid.resource
 
 import android.accounts.Account
-import android.annotation.SuppressLint
 import android.content.ContentProviderClient
 import android.content.ContentValues
-import android.content.Context
 import at.bitfire.davdroid.db.SyncState
 import at.bitfire.ical4android.DmfsTaskList
 import at.bitfire.ical4android.DmfsTaskListFactory
@@ -29,24 +27,6 @@ class LocalTaskList private constructor(
         id: Long
 ): DmfsTaskList<LocalTask>(account, provider, providerName, LocalTask.Factory, id), LocalCollection<LocalTask> {
 
-    companion object {
-
-        @SuppressLint("Recycle")
-        @Throws(Exception::class)
-        fun onRenameAccount(context: Context, oldName: String, newName: String) {
-            TaskProvider.acquire(context)?.use { provider ->
-                val values = ContentValues(1)
-                values.put(Tasks.ACCOUNT_NAME, newName)
-                provider.client.update(
-                        Tasks.getContentUri(provider.name.authority),
-                        values,
-                        "${Tasks.ACCOUNT_NAME}=?", arrayOf(oldName)
-                )
-            }
-        }
-
-    }
-    
     private val logger = Logger.getGlobal()
 
     private var accessLevel: Int = TaskListColumns.ACCESS_LEVEL_UNDEFINED
